+++ CHANGES.md
@@ -1,6 +1,6 @@
 #### [unreleased]
-* run GitHub Updater always but only show Settings for privileged users
 * run API calls for everyone, not just privileged users, hopefully this allows for better integration with remote management services
+* only show Settings for privileged users
 
 #### 9.0.1 / 2019-12-04
 * fix PHP version check, fixes [#824](https://github.com/afragen/github-updater/issues/824)