+++ CHANGES.md
@@ -1,6 +1,6 @@
 #### [unreleased]
 * run API calls for everyone with wp-cron, not just privileged users, hopefully this allows for better integration with remote management services
-* don't run API calls for non-privileged users when bypassing wp-cron
+* don't run API calls on frontend when bypassing wp-cron
 * only show Settings for privileged users
 
 #### 9.0.1 / 2019-12-04